refactor: rename Button component
